use table_schema isntead of schema_name with crate >= 0.57
